requirements.txt
streamlit
google-generativeai
sentence-transformers
faiss-cpu
pandas

numpy
yfinance
plotly
matplotlib
PyPDF2
python-dotenv
reportlab
requests
regex
scikit-learn
pytest
pytest-cov
numba
nltk
loguru
